+++ app/src/c/game/game.h
@@ -5,4 +5,4 @@ void game_init(State* stateRef);
 
 void game_update_stats(short identifier);
 
-void game_add_monster(ResourceValue monster, int power);
+void game_add_monster(ResourceValue monster, short power);
